+++ contracts/sales/presets/FixedPriceRefundSeller.sol
@@ -34,7 +34,7 @@ contract FixedPriceRefundSeller is
 
     function _beforePurchase(
         address to,
-        uint256 num,
+        uint64 num,
         uint256 cost
     )
         internal
@@ -42,7 +42,7 @@ contract FixedPriceRefundSeller is
         override(FixedSupplyRefund, InternalCostSeller)
         returns (
             address,
-            uint256,
+            uint64,
             uint256
         )
     {
@@ -52,13 +52,13 @@ contract FixedPriceRefundSeller is
 
     function _afterPurchase(
         address to,
-        uint256 num,
+        uint64 num,
         uint256 cost
     ) internal virtual override(FixedSupplyRefund, Seller) {
         CappedRefund._afterPurchase(to, num, cost);
     }
 
-    function purchase(address to, uint256 num) external payable whenNotPaused {
+    function purchase(address to, uint64 num) external payable whenNotPaused {
         _purchase(to, num, 0);
     }
 }
